+++ include/uit/countdown/Timer.hpp
@@ -4,7 +4,7 @@
 #include <chrono>
 #include <stddef.h>
 
-#include "../utility/chrono_utils.hpp"
+#include "../chrono/chrono_utils.hpp"
 
 #include "impl/CountdownIterator.hpp"
